refactor: moving containers folders to post feature
