+++ src/features/Posts/containers/SubmitPost/index.tsx
@@ -1,9 +1,9 @@
 import { useState, ChangeEvent } from "react";
-import { Button } from "../../../components/Button";
-import { Cards } from "../../../components/Cards";
-import { ImageInput } from "../../../components/ImageInput";
-import { Input } from "../../../components/Input";
-import { Textarea } from "../../../components/Textarea";
+import { Button } from "../../../../components/Button";
+import { Cards } from "../../../../components/Cards";
+import { ImageInput } from "../../../../components/ImageInput";
+import { Input } from "../../../../components/Input";
+import { Textarea } from "../../../../components/Textarea";
 import { FormContainer } from "./style";
 
 export function SubmitPost() {
