+++ src/features/Posts/index.tsx
@@ -1,5 +1,5 @@
 import { Page } from "../../components/Page";
-import { SubmitPost } from "../containers/SubmitPost";
+import { SubmitPost } from "./containers/SubmitPost";
 
 export function Posts() {
   return (
